Move the states generator to the states module
